Move token to variable logic to own file
